Replace "Pubsub Subscription Different Project" example with making all examples cross-project-friendly (#9637) (#1794)
* Fix "Pubsub Subscription Different Project" example

PR #2342 changed this example from using `id` to `name`, which removed the project ID from the topic. However, the point of this example is to show how to set up a cross-project subscription, so the example no longer works. This change reverts that change from PR #2342.

Fixes hashicorp/terraform-provider-google#11642.
Fixes hashicorp/terraform-provider-google#6024.

* Make all examples cross-project-friendly and update topic description

* Fix typo

---------


[upstream:b3ae8dcd7e476f796a8b1ba67ce63cea711be127]

Signed-off-by: Modular Magician <magic-modules@google.com>