Cai2Hcl: refactor of converter_map and split of utils.go (#9378)
* Add cai2hcl generated converters for 3 Compute resources.

These are: BackendService, GlobalBackendService, ForwardingRule.

This is the output of cai2hcl provider, commited as is
without the generator Ruby code itself, as agreed with
library owners.
Reason: limited capacity of the supporting team to
perform code reviews and the future need to port this
solution to golang as a part of migration to go. As soon
the staffing and/or migration is done, we will come back
to implement the generated converters as
originally planned.

* Split current PR into 3 chunks:
1. Refactor of shared files (Uber Converter)
2. Add ForwardingRuleConverter (generated)
3. Add BackendServiceConverters (generated)
+ ability to match converters by name

This commit removes part2 and part3, leaving only part1 in this branch.

* Fix versions after merge conflict

* Remove auto-generated code header from FR

* Remove unused util function

* Fix merge conflict resolution issue

* Remove unnecessary result map normalization (will needed later with generated converters)

* Simplify cai2hcl by reworking UberConverter's as a single ConverterMap.

1. Rename UberConverter to ConverterMap.
2. Remove ConverterMap's from each API folder and replace them with a single top-level ConverterMap

* Simplify converters

* Flatten converter_map.go and use strings as converter names instead of constants.
[upstream:ddc054c826c23b8d0bf39273fb71c4c32b038740]

Signed-off-by: Modular Magician <magic-modules@google.com>