Fix Spanner update when adding/removing autoscaling config (#9814) (#1875)
* Add reproducer test for error.

Creating a Spanner instance without auto-scaling and then updating it to
add auto-scaling should work, but currently doesn't.

* Fix `fieldMask` for spanner updates when enabling/disable autoscaling.

When adding/removing auto-scaling config wholesale, the API expects only
the top-level `autoscalingConfig` key to be listed in the `fieldMask`.
Listing sub-fields of the autoscaling config causes an error.

* Add tests for dropping auto-scaling.

Not sure if this was broken before, but makes sense to me that we should
test in both directions.

* Apply review feedback

* Fix indentation in template
* Rename test instances to match expected pattern
* Consolidate tests

* Fix name-too-long errors
[upstream:8f488115df8493669aae88a65b80a2fcb31d7557]

Signed-off-by: Modular Magician <magic-modules@google.com>